+++ src/v1.x/packages/a2ui-renderer/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @copilotkit/a2ui-renderer
 
+## 1.51.3-next.6
+
+### Patch Changes
+
+- @copilotkit/react-core@1.51.3-next.6
+
+## 1.51.3-next.5
+
+### Patch Changes
+
+- Updated dependencies [e89fe98]
+  - @copilotkit/react-core@1.51.3-next.5
+
+## 1.51.3-next.4
+
+### Patch Changes
+
+- @copilotkit/react-core@1.51.3-next.4
+
 ## 1.51.3-next.3
 
 ### Patch Changes