+++ src/v1.x/packages/react-core/CHANGELOG.md
@@ -1,5 +1,33 @@
 # ui
 
+## 1.51.3-next.6
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.51.3-next.6
+- @copilotkit/shared@1.51.3-next.6
+- @copilotkitnext/core@1.51.3-next.6
+- @copilotkitnext/react@1.51.3-next.6
+
+## 1.51.3-next.5
+
+### Patch Changes
+
+- e89fe98: fix(v1.x): address issue causing onSubmitMessage to fire after agent response
+  - @copilotkit/runtime-client-gql@1.51.3-next.5
+  - @copilotkit/shared@1.51.3-next.5
+  - @copilotkitnext/core@1.51.3-next.5
+  - @copilotkitnext/react@1.51.3-next.5
+
+## 1.51.3-next.4
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.51.3-next.4
+- @copilotkit/shared@1.51.3-next.4
+- @copilotkitnext/core@1.51.3-next.4
+- @copilotkitnext/react@1.51.3-next.4
+
 ## 1.51.3-next.3
 
 ### Patch Changes