+++ src/v1.x/packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @copilotkit/shared
 
+## 1.51.3-next.6
+
+## 1.51.3-next.5
+
+## 1.51.3-next.4
+
 ## 1.51.3-next.3
 
 ### Patch Changes