+++ src/v2.x/packages/sqlite-runner/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @copilotkitnext/sqlite-runner
 
+## 1.51.3-next.6
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.51.3-next.6
+
+## 1.51.3-next.5
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.51.3-next.5
+
+## 1.51.3-next.4
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.51.3-next.4
+
 ## 1.51.3-next.3
 
 ### Patch Changes